requirements.txt
pytubefix
pyTelegramBotAPI
python-dotenv